Convert build to gradle.
