+++ src/main/java/com/dumbster/smtp/MailMessage.java
