+++ src/main/java/com/dumbster/smtp/mailstores/RollingMailStore.java
