+++ src/test/java/com/dumbster/smtp/MailMessageTest.java
