+++ src/test/java/com/dumbster/smtp/RollingMailStoreTest.java
