+++ src/test/java/com/dumbster/smtp/ServerOptionsTest.java
