+++ src/test/java/com/dumbster/smtp/StringUtil.java
